Store a "std::pair<Key, T>" instead of a "std::pair<const Key, T>".
We now return a "const std::pair<Key, T>&" in the iterators instead of "std::pair<const Key, T>&". The only way to modify the value in the map is by calling something like it.value()=42.
This breaks compatibility with the std::unordred_map interface, but allow us to speed-up insert and rehash by being able to move the key of std::pair.